+++ src/initialData/index.ts
@@ -2,4 +2,4 @@ export { getOperationalStates, getOperationalErrorState } from './getOperational
 export { getSupportedCleanModes } from './getSupportedCleanModes.js';
 export { getSupportedAreas } from './getSupportedAreas.js';
 export { getBatteryStatus, getBatteryState } from './getBatteryStatus.js';
-export { getSupportedScenes } from './getSupportedScenes.js';
+export { getSupportedRoutines } from './getSupportedRoutines.js';